Orchestrator/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon
apache-airflow-providers-postgres
apache-airflow-providers-common-sql
astronomer-cosmos
astro-sdk-python
boto3
sqlalchemy
psycopg2-binary
asyncpg 